Further development of approach and landing algorithms, in particular consistent use of metres for ground distances, and additional features to improve robustness.
LatitudeSmoothed and LongitudeSmoothed no longer test for ILS Range; this is moved into the adjust_track function to cater for flights with ILS and non-ILS approaches.

KPV: AltitudeAtSuspectedLevelBust extensively reworked, and may still need further improvement.

KTI: LowestPointOnApproach reinstated as this is needed to trigger other KPVs for the approach phase.

Library: slices_not now returns the entire scope as a slice in the absence of any slices in the list.