Frame specific derived parameters extended where necessary for 737-1 and 737-i frames.
GearExtending and GearRetracting flight phases reordered to make monitored case the first option and the backstop of guessing 5 seconds for the gear to travel the second.

TCAS modes have an indexing error corrected and a trap for possible all-invalid accelerometer data added (this trap is probably not required for correct indexing, but was added part way through development and does no harm).