KTI test corrections.
Takeoff acceleration start algorithm was changed to cater for masked data at low speed, so the test data needed to be updated to reflect this condition.

Touchdown instant was being computed incorrectly because the sample data arrays were integer, causing the smoothed height computation to work in integer units only. Now adjusted to be float.