Vertical Speed Inertial test code relocated
A short piece of test code which allows the parameters used to detect
touchdown points in the absence of a reliable Gear On Ground signal has
been moved to the procedure touchdown_inertial where it rightfully
belongs. This was used when debugging the 737-3A data frame which has a
recorded inertial vertical speed which is not consistent with the
equivalent POLARIS parameter, so is not used for data processing.