Ground track precise reverts to dead reckoning
If there are fewer than two straight legs during taxi out, the
optimisation algorithm is inhibited. In these cases it is better to draw
a ground track using dead reckoning (heading and groundspeed) than to
use the prepared latitude and longitude track. As we are going to revise
this algorithm, the test thresholds have been set wide to ensure the
test passes at this time.