Curving visual approaches corrected.
When computing the track for a curving visual approach, the assumption
that the aircraft is following the runway centreline is clearly
incorrect, so the algorithm has been improved to take account only of
the touchdown point and put the track onto the runway at that location.
The curving descent path is then just the smoothed latitude and
longitude.