Increased frequency of Daylight to 0.25 for LFLs without superframes.
Parameters from an LFL without a superframe were failing to align to Daylight with a frequency of 1/64.  This is because we do not split on Superframe boundaries if the frame does not include it. The downside of the fix is that we will be outputting Daylight at a higher frequency than required.