More step_values devel & minor fixes
step_values is still being improved, and FlapLeverSynthetic has been
written to study this, but the current version is not fit for production
use. See comments in the code.

blend_parameters has a problem with downsampling due to reductions in
sample rate with valid slices that may have odd sample numbers. For the
interim radio altitude and ILS glideslope (which use this function)
return the highest sample rate parameter.

Track computations in ApproachRange have been moved to ensure the valid
count (np.ma.count) applies only to the approach in question and not to
the complete flight.