test: kpv: fix tests broken by 11ef0d5
All tests for the For X Sec KPVs were broken due to additional
dependencies that broke the can_operate() test and use of an additional
function for the second argument passed to the create_kpv_* helper
functions. As a result the mock tests have had to be removed as they
currently cannot support the level of complexity introduced.