Prevent interpolation in some flap-related KPVs.
KPVs that have a value that should be a flap detent should not have
interpolated values when looking up based on a KTI with a floating point
index. Added support for optionally disabling interpolation in the
library routine value_at_index() and the ability to pass the flag
through to it via create_kpvs_at_ktis(). Updated with extra tests.