Fix tests related to v-speed derived parameters.
Updated NodeTest.test_can_operate() to allow us to check for only the
important combinations. This still needs to be refactored out somewhere
so that we can use the same version in all test modules. We also need to
eliminate the length-only test as it is pointless.

Fixed the operational combination tests for v-speed related derived
parameters and also one of the derive tests due to lookup table fixes in
FlightDataUtilities commit b8e0ece.