ClimbCruiseDescent squashed above 10k ft.
This chance replaces AltitudeAAL with AltitudeSTD and compresses the
heights above 10.000ft STD so that in the cruise the height changes
required to initiate a new phase are five times greater. In this way,
small changes close to the airfield are still recognised as a new
climb/descent period, while at altitude similar height changes are
ignored.