gear travel update and taxi phases reduced
Gear Extending and Retracting phases use red warnings, but sometimes one
of these is not available. Phases reworked to not rely upon all being
available.
Taxi phases pick up the Mobile duration to limit the time on stand, and
hence improve the quality of data monitoring for the taxi phases.